Reorg new functions, add docs
