+++ packages/sdk-router/src/operations/index.ts
@@ -1,4 +1,3 @@
 export * from './bridge'
 export * from './pools'
 export * from './swap'
-export * from './query'
