+++ packages/sdk-router/src/operations/swap.ts
@@ -2,7 +2,7 @@ import { PopulatedTransaction } from '@ethersproject/contracts'
 import { BigNumber } from '@ethersproject/bignumber'
 
 import { BigintIsh } from '../constants'
-import { Query, SwapQuote } from '../module'
+import { Query, SwapQuote, applySlippage } from '../module'
 import { handleNativeToken } from '../utils/handleNativeToken'
 import { SynapseSDK } from '../sdk'
 import { TEN_MINUTES, applyOptionalDeadline } from '../utils/deadlines'
@@ -63,4 +63,20 @@ export async function swapQuote(
   }
   query.deadline = applyOptionalDeadline(deadline, TEN_MINUTES)
   return { routerAddress, maxAmountOut, query }
+}
+
+/**
+ * Applies slippage to the given swap query.
+ *
+ * @param queryPrecise - The swap query, coming from `swapQuote()`
+ * @param slipNumerator - The numerator of the slippage percentage
+ * @param slipDenominator - The denominator of the slippage percentage, defaults to 10000
+ * @returns The swap query with slippage applied
+ */
+export const applySwapSlippage = (
+  queryPrecise: Query,
+  slipNumerator: number,
+  slipDenominator: number = 10000
+): Query => {
+  return applySlippage(queryPrecise, slipNumerator, slipDenominator)
 }
